Fix two CI failures on the wrapper review branch
- read_geometry_data's nanobind binding now returns a variant-typed
  object (to avoid a bad_variant_access on float32 geometry), so its
  Python wrapper's return type was no longer accurate; widen it to
  float32 | float64 and cast to float64 in read_mesh, which only ever
  builds a float64 CoordinateElement.
- static_cast PetscErrorCode to int before formatting it with spdlog,
  matching common::petsc::error's existing convention; newer fmt
  releases (as used by the macOS Homebrew CI job) refuse to format an
  opaque PetscErrorCode directly.

+++ python/dolfinx/wrappers/dolfinx_wrappers/petsc.h
@@ -76,7 +76,8 @@ void declare_petsc_discrete_operators(nb::module_& m)
           // Destroy A before rethrowing. Report rather than throw on
           // failure: throwing here would replace the in-flight exception.
           if (PetscErrorCode ierr = MatDestroy(&A); ierr != 0)
-            spdlog::error("MatDestroy failed with error code {}.", ierr);
+            spdlog::error("MatDestroy failed with error code {}.",
+                          static_cast<int>(ierr));
           throw;
         }
         return A;
@@ -108,7 +109,8 @@ void declare_petsc_discrete_operators(nb::module_& m)
           // Destroy A before rethrowing. Report rather than throw on
           // failure: throwing here would replace the in-flight exception.
           if (PetscErrorCode ierr = MatDestroy(&A); ierr != 0)
-            spdlog::error("MatDestroy failed with error code {}.", ierr);
+            spdlog::error("MatDestroy failed with error code {}.",
+                          static_cast<int>(ierr));
           throw;
         }
         return A;
@@ -137,7 +139,8 @@ void declare_petsc_discrete_operators(nb::module_& m)
           // Destroy A before rethrowing. Report rather than throw on
           // failure: throwing here would replace the in-flight exception.
           if (PetscErrorCode ierr = MatDestroy(&A); ierr != 0)
-            spdlog::error("MatDestroy failed with error code {}.", ierr);
+            spdlog::error("MatDestroy failed with error code {}.",
+                          static_cast<int>(ierr));
           throw;
         }
         return A;
